Enh#98 (#99)
* Moved prwlr.core.read_profiles from setUp to the test function. Moved import as well.

* Moved prwlr.core.read_network from setUp to the test function. Moved import as well.

* tearDown checks if file exists before removing it. prwlr.core.save_profiles moved from setUp to the test function

+++ tests/tests.py
@@ -677,38 +677,26 @@ class CoreTests(unittest.TestCase):
             }]
         )
 
-        self.test_profiles_srs = prwlr.core.read_profiles(
-            self.test_profiles_filename,
-            index_col=[0],
-        )
-        self.test_network_df = prwlr.core.read_network(
-            self.test_network_filename,
-            index_col=[0],
-        )
-
-        prwlr.core.save_profiles(
-            self.ref_profiles_srs,
-            self.test_saved_profiles_filename,
-        )
-        self.test_profiles_after_save_srs = prwlr.core.read_profiles(
-            self.test_saved_profiles_filename,
-            index_col=[0],
-        )
-
     def tearDown(self):
         """
         Removes files created during the tests.
         """
-        os.remove(self.test_saved_profiles_filename)
+        if os.path.exists(self.test_saved_profiles_filename):
+            os.remove(self.test_saved_profiles_filename)
 
     def test_read_profiles(self):
         """
         Tests if prwlr.core.read_profiles returns pandas.Series with correct
         prwlr.profiles.Profile objects.
         """
+        import prwlr.core
+
         pd.testing.assert_series_equal(
             self.ref_profiles_srs,
-            self.test_profiles_srs,
+            prwlr.core.read_profiles(
+                self.test_profiles_filename,
+                index_col=[0],
+            ),
         )
 
     def test_read_network(self):
@@ -716,9 +704,14 @@ class CoreTests(unittest.TestCase):
         Tests if prwlr.core.read_network returns pandas.DataFrame with correct
         prwlr.profiles.Profile objects.
         """
+        import prwlr.core
+
         pd.testing.assert_frame_equal(
             self.ref_network_df,
-            self.test_network_df,
+            prwlr.core.read_network(
+                self.test_network_filename,
+                index_col=[0],
+            ),
             check_like=True,
         )
 
@@ -726,9 +719,18 @@ class CoreTests(unittest.TestCase):
         """
         Tests if prwlr.core.save_profiles writes CSV file with correct values.
         """
+        import prwlr.core
+
+        prwlr.core.save_profiles(
+            self.ref_profiles_srs,
+            self.test_saved_profiles_filename,
+        )
         pd.testing.assert_series_equal(
             self.ref_profiles_srs,
-            self.test_profiles_after_save_srs,
+            prwlr.core.read_profiles(
+                self.test_saved_profiles_filename,
+                index_col=[0],
+            ),
         )
